コンポーネントWebsiteSettingsの切り出し
