+++ dashboard/src/components/Input.tsx
@@ -1,0 +1,32 @@
+import { styled } from '@macaron-css/solid'
+import { vars } from '/@/theme'
+
+export const InputLabel = styled('div', {
+  base: {
+    fontSize: '16px',
+    alignItems: 'center',
+    fontWeight: 700,
+    color: vars.text.black1,
+
+    marginBottom: '4px',
+  },
+})
+
+export const InputBar = styled('input', {
+  base: {
+    padding: '8px 12px',
+    borderRadius: '4px',
+    border: `1px solid ${vars.bg.white4}`,
+    fontSize: '14px',
+    marginLeft: '4px',
+
+    width: '320px',
+
+    display: 'flex',
+    flexDirection: 'column',
+
+    '::placeholder': {
+      color: vars.text.black3,
+    },
+  },
+})
